Changed the file structure.
This commit changes the file structure of the-mac Tool-Kit to move all
Android development projects to an Android Projects folder, all iOS
projects to an iOS Projects folder, and all Windows Projects to a
Windows Projects folder. Additionally, Apple sample code has been added
to the iOS projects folder for the use of all iOS developers. This
sample content includes but is not limited to adding iAD content to
applications in various ways.